requirements.txt
anthropic>=0.7.0
typer[all]>=0.9.0
rich>=13.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
beautifulsoup4>=4.12.0
stripe>=7.0.0
flask>=2.3.0
fastapi>=0.104.0
uvicorn>=0.24.0
requests>=2.31.0